requirements.txt
groq
google-cloud-texttospeech==2.4.0
google-cloud-speech==2.4.0
google-cloud-storage==2.4.0
moviepy==1.0.3
requests==2.26.0
google-api-python-client==2.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
python-dotenv==1.0.1
pytubefix
bs4
pillow==9.5.0
ffmpeg-python
pytz
flask # used for tiktok automation which is not included in repo
waitress # used for tiktok automation which is not included in repo
